Resolved token swap performance problem one more time and started implementing topic aware conversations.
