+++ README.md
@@ -68,9 +68,9 @@ A modular, production-ready dialogue summarization system built with PyTorch Lig
     
     # For prediction
     python scripts/inference.py predict \
-    /home/wb2x/workspace/dialogue-summarizer/outputs/models/best-epoch=00-val_rouge_f=val/rouge_f=0.6014.ckpt \
+    /home/wb2x/workspace/dialogue-summarizer/outputs/models/best-epoch=01-val_rouge_f=val/rouge_f=0.6217.ckpt \
     data/test.csv \
-    subission_R060.csv
+    subission_unbiased.csv
 
     # (DEPRECATED) For prediction (before using centralized config)
     python scripts/inference.py predict \
