+++ configs/postprocessing/swap_regular_names.yaml
@@ -19,13 +19,7 @@ remove_tokens:
   - "</s>"
   - "<pad>"
   - "<unk>"
-  # - "민준"
-  # - "서연"
-  # - "지훈"
-  # - "수진"
-  # - "현우"
-  # - "은영"
-  # - "동현"
+  
 text_cleaning:
   strip_whitespace: true
   normalize_whitespace: true
